refactor: moved slide data to home page domain
