+++ src/domains/store/homePage/constants/index.ts
@@ -1,2 +1,3 @@
 export * from "./wideCards";
 export * from "./collections";
+export * from "./homepageData";
